+++ src/VaultSharp/V1/SecretsEngines/PKI/PKISecretsEngineProvider.cs
@@ -93,5 +93,12 @@ namespace VaultSharp.V1.SecretsEngines.PKI
 
             return result;
         }
+
+        public async Task<Secret<CertificateKeys>> ListRevokedCertificatesAsync(string pkiBackendMountPoint = null)
+        {
+            var result = await _polymath.MakeVaultApiRequest<Secret<CertificateKeys>>(pkiBackendMountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.PKI, "/certs/revoked?list=true", HttpMethod.Get).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+
+            return result;
+        }
     }
 }